requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
idna==3.6
python-dotenv==1.0.0
requests==2.31.0
selectolax==0.3.17
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
